+++ config/bootstrap.php
@@ -37,16 +37,21 @@ if (!Configure::check(THUMBER . '.target')) {
     Configure::write(THUMBER . '.target', TMP . 'thumbs');
 }
 
-if (!in_array(Configure::read(THUMBER . '.driver'), ['imagick', 'gd'])) {
-    trigger_error(
-        sprintf('The driver %s is not supported', Configure::read(THUMBER . '.driver')),
-        E_USER_ERROR
-    );
+//Checks for driver
+$driver = Configure::read(THUMBER . '.driver');
+
+if (!in_array($driver, ['imagick', 'gd'])) {
+    trigger_error(sprintf('The driver %s is not supported', $driver), E_USER_ERROR);
 }
 
-if (!is_writeable(Configure::read(THUMBER . '.target'))) {
-    trigger_error(
-        sprintf('Directory %s not writeable', Configure::read(THUMBER . '.target')),
-        E_USER_ERROR
-    );
+//Checks for target directory
+$target = Configure::read(THUMBER . '.target');
+
+if (!file_exists($target)) {
+    //@codingStandardsIgnoreLine
+    @mkdir($target, 0777, true);
+}
+
+if (!is_writeable($target)) {
+    trigger_error(sprintf('Directory %s not writeable', $target), E_USER_ERROR);
 }