+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.x branch
 ## 1.0 branch
 ### 1.0.5
+* the target directory is created automatically, if it does not exist;
 * added `THUMBER` constant.
 
 ### 1.0.4